Add namespaces for parent classes.
This is useful when a derived object (such as WinSystem) uses properties
from a parent object (i.e. System). Previously, the System namespace
would not have been added.

Ultimately, we need to add namespaces to TypedFields as well as the Entity
classes themselves, but this probably won't happen until we split off the
"xs" library.