Fixed to actually test the proper wide character strings under python2.
The code
  ustr = six.u("<some utf8 bytes>")

doesn't do what the original author thought it did under python2.
That's a narrow char string of utf8 bytes, and six.u() is hard-coded
to decode using the "unicode_escape" encoding, which is the "\u1234"
syntax.  There are no unicode escapes there, so a proper decoding
never happens.  I fixed it to use unicode escapes instead of utf8
encoded literals.